requirements.txt
altair==5.2.0
asttokens==2.0.8
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.1
gitdb==4.0.11
GitPython==3.1.40
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-core==4.11.1
jupyter_client==7.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.6
numpy==1.24.4
opencv-python==4.8.1.78
packaging==21.3
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.31
protobuf==4.25.1
psutil==5.9.2
pure-eval==0.2.2
pyarrow==14.0.2
pydeck==0.8.1b0
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==24.0.1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.15.2
six==1.16.0
smmap==5.0.1
stack-data==0.5.1
streamlit==1.29.0
streamlit-cropper==0.2.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.4.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.5
zipp==3.17.0
